- minor change to ImplShiTomasi where an unnecessary conversion from double to float was removed. - FastHessianFeatureDetector wrapper returns the correct radius now

+++ main/feature/src/boofcv/alg/feature/detect/intensity/impl/ImplShiTomasiCornerWeighted_F32.java
@@ -41,9 +41,9 @@ public class ImplShiTomasiCornerWeighted_F32 extends ImplSsdCornerWeighted_F32
 		// compute the smallest eigenvalue
 		float left = (totalXX + totalYY) * 0.5f;
 		float b = (totalXX - totalYY) * 0.5f;
-		double right = Math.sqrt(b * b + totalXY * totalXY);
+		float right = (float)Math.sqrt(b * b + totalXY * totalXY);
 		
 		// the smallest eigenvalue will be minus the right side
-		return (float)(left - right);
+		return left - right;
 	}
 }
